# infrastructure/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Fund Transfer System API
  description: |
    Multi-service PHP/Symfony fund transfer platform.
    All endpoints are accessible through NGINX on port 80.
    Each service is routed by path prefix.
  version: "1.0.0"
  contact:
    name: Fund Transfer System
servers:
  - url: http://localhost
    description: Local Docker environment (via NGINX)
...
security:
  - BearerAuth: []
paths:
...
  /account:
    post:
      tags: [Accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
            example:
              userId: 1
              balance: "1000.00"
              currency: USD
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /account/{id}:
    get:
      tags: [Accounts]
      summary: Get an account by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /account/{id}/debit:
    post:
      tags: [Accounts]
      summary: Debit (subtract) an amount from an account
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
            example:
              amount: "50.00"
      responses:
        "200":
          description: Account debited, updated account returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /account/{id}/credit:
    post:
      tags: [Accounts]
      summary: Credit (add) an amount to an account
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
            example:
              amount: "200.00"
      responses:
        "200":
          description: Account credited, updated account returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Transactions ──────────────────────────────────────────────────────────
...
  /ledger/entry:
    post:
      tags: [Ledger]
      summary: Record a new ledger entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerEntryRequest"
            example:
              transactionId: 1
              accountId: 2
              entryType: debit
              amount: "100.00"
      responses:
        "201":
          description: Ledger entry recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /ledger/account/{accountId}:
    get:
      tags: [Ledger]
      summary: Get all ledger entries for an account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Array of ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
        "500":
          $ref: "#/components/responses/ServerError"
  /ledger/transaction/{transactionId}:
    get:
      tags: [Ledger]
      summary: Get all ledger entries for a transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Array of ledger entries (both legs of the transfer)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
        "500":
          $ref: "#/components/responses/ServerError"

# ── Components ───────────────────────────────────────────────────────────────
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
      required: [error, code]
...
    CreateAccountRequest:
      type: object
      required: [userId, balance, currency]
      properties:
        userId:
          type: integer
          example: 1
        balance:
          type: string
          example: "1000.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
    Account:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
        balance:
          type: string
          example: "1000.00"
        currency:
          type: string
          example: USD
        createdAt:
          type: string
          format: date-time
    AmountRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: string
          example: "50.00"
...
    CreateLedgerEntryRequest:
      type: object
      required: [transactionId, accountId, entryType, amount]
      properties:
        transactionId:
          type: integer
          example: 1
        accountId:
          type: integer
          example: 2
        entryType:
          type: string
          enum: [debit, credit]
          example: debit
        amount:
          type: string
          example: "100.00"
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
          example: 1
        transactionId:
          type: integer
          example: 1
        accountId:
          type: integer
          example: 2
        entryType:
          type: string
          enum: [debit, credit]
          example: debit
        amount:
          type: string
          example: "100.00"
        createdAt:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Invalid JSON body.
            code: 400
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Resource not found.
            code: 404
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: An unexpected error occurred.
            code: 500